refactor(LOTTO): earningMap 제거 및 class 경로 변경
