+++ src/main/java/lotto/util/RandomNumberUtils.java
@@ -1,7 +1,7 @@
 package lotto.util;
 
-import lotto.LottoNumber;
-import lotto.LottoNumbers;
+import lotto.domain.LottoNumber;
+import lotto.domain.LottoNumbers;
 
 import java.util.ArrayList;
 import java.util.Collections;
