+++ src/test/java/lotto/domain/LottoGameTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.domain;
 
 import org.junit.jupiter.api.Test;
